[REF] pos_customer_wallet: Refactor to .esm.js
